Score system hook up ui to score

+++ Assets/Code/Score/ScoreModel.cs
@@ -5,6 +5,6 @@ public class ScoreModel
 {
     public float InvasionScore { get; set; }
     public int DogsDead { get; set; }
-    public long Score { get; set; }
+    public int Score { get; set; }
     public float Time { get; set; }
 }
